chore(security): strip dead plugin orchestrators (#116) + drop cve-lite self-link (#117)
#116 — applyInstallGate (apply-install-gate.ts) and runAllPluginCards
(runner.ts) were scaffolding-only. /update hand-rolls a minimal single-plugin
binary-swap and never calls processResult; the plugins GET route is host-only
(no project context); and the accordion builds PluginCardEntry[] client-side
from the per-plugin status endpoint — so neither orchestrator had a prod caller
or a near-term one. Removed both modules + their tests. Preserved the live UI
contract by moving the PluginCardEntry type into types.ts and repointing its
five importers. Richer multi-plugin chaining + processResult block-parsing stays
the #116 follow-up (design in git history + spec), to build when a second
installGate plugin or block-parsing UX actually lands.

#117 — dropped sourceDeepLinks={{ 'cve-lite': '/security' }} from the security
accordion; the cve-lite card's "Open dashboard" linked back to the same page.
SourceCard gates the link on deepLinkHref, so removing the prop simply hides it;
the optional prop stays supported for future legitimate cross-page links.

Verified: typecheck clean, 215 tests pass, build green.
